move common db methods to shared module
